First pass at Plugin config forms.
Added display area for plugin config forms at the bottom of the Admin's Plugins page.  This page will need to be restyled a bit to accommodate.

Added a POST handler for the admin to accept plugin form submissions and pass them on to the plugin's code for displaying forms.

The FormUI class is the base form class for adding new forms to the Plugins admin page.  Included in that same file are the FormControl class, a base class for deriving form controls, and FromControlText, which is for creating text controls inside of FormUI.  These extra classes should probably be moved to their own files.

Also included minor bugfixes for Utils::firedebug().

git-svn-id: http://svn.habariproject.org/habari/trunk@695 653ae4dd-d31e-0410-96ef-6bf7bf53c507