Hopefully, the final commits on plugin/XMLRPC interaction tonight.
You should now be able to create a function in a plugin class that is prefixed with "xmlrpc_" and have it automatically register to handle the XMLRPC function of that name.  A double-underscore in the function name will be converted to a dot.

A new function in Plugins, Plugins::xmlrpc() handles calling XMLRPC hook functions.  It returns the value from the first matching registered function, rather than going through every defined function and filtering.  This value is converted into XML and returned to the remote caller.

git-svn-id: http://svn.habariproject.org/habari/trunk@696 653ae4dd-d31e-0410-96ef-6bf7bf53c507