Adding a filter call to admin/publish.php to allow plugins to manipulate the list of post statuses that a user may select when composing or editing a post.
Updating the undelete plugin to make use of this new filter.  The upshot 
is that "deleted" is no longer a status one may manually select when 
composing or editing a post.

The filter naming convention I used for this is 
"section_subsection_item".  In this case, it's 
"admin_publish_list_post_statuses".  The intent it to make it clear when 
reading the code where this filter is being triggered.  This might not 
be applicable in all instances.  More or less granularity is an item for 
discussion.



git-svn-id: http://svn.habariproject.org/habari/trunk@765 653ae4dd-d31e-0410-96ef-6bf7bf53c507