This will be the last revision of the AtomHandler until Habari can handle media.
I added the element <accept> to make it explicit that we only support atom entries.

Also fixed a bug that passed under the radar, the <link> for previous and next atom feed were wrong, it should now work as expected.

And of course, added the app:edited which makes the APE successfully finish, although it shows an error for multi-post entries, that is a bug in the APE code, will have to report it to Tim Bray. (I may be wrong, but look at my code, and at the output, dates are ordered properly and the APE does not return an error for a specific entry.)

git-svn-id: http://svn.habariproject.org/habari/trunk@805 653ae4dd-d31e-0410-96ef-6bf7bf53c507