Minor updates to the undelete plugin.
Upon activation, it will register a new post status named "deleted".  This status is marked as "internal", and should therefore be suppressed from display to the end user in most circumstances.

Also applies some FormUI to permit the user to confgure the CSS applied when viewing deleted posts.  This CSS is only output in the <head> element for logged-inusers, since anonymous users should never see deleted posts anyway.  A default value of red background with strikethrough text is defined.



git-svn-id: http://svn.habariproject.org/habari/trunk@898 653ae4dd-d31e-0410-96ef-6bf7bf53c507