Plugins have access to select boxes via FormUI.  Use like this:
$twitter_post = $ui->add('select', 'post_status', 'Autopost to Twitter:');
$twitter_post->options = array('0' => 'Disabled', '1' => 'Enabled');

The value that matches the option setting will automatically be selected.

Added a new password type field to FormUI.  Values that are displayed in a password field are actually the md5 of the field value.  If the field value is changed, the new value is stored.  If it remains the same, then the old value is not overwritten with its md5 counterpart.  This prevents passwords stored in the options table from being written in the raw into password fields in the FormUI.


git-svn-id: http://svn.habariproject.org/habari/trunk@993 653ae4dd-d31e-0410-96ef-6bf7bf53c507