Make publish page settings area pluggable.
Update admin.css to use generated class name for tags tab.  Display tabs from array on publish.php.  Fetch tab contents for display in adminhandler.php.  Make fetch() return the output buffer and clean it.  Make the Theme class in theme.php support fetch() in addition to display().

Use the filter 'publish_controls' to filter the array sent to the publish page.  The key is the tab name, the value is what is displayed in the tab.

git-svn-id: http://svn.habariproject.org/habari/trunk@1209 653ae4dd-d31e-0410-96ef-6bf7bf53c507