$post->tags should now return array() if there are no tags.
XMLRPC does not like '', because it will not send an empty array struct, which it should. (metaWeblog is an example - grrr)

It will not break your empty() conditions, it should strictly fix XMLRPC bugs or unexpected behaviors.

Thanks to MarsEdit's dcj (dev) for helping out on this one! (great software, shameless plug!)

git-svn-id: http://svn.habariproject.org/habari/trunk@1462 653ae4dd-d31e-0410-96ef-6bf7bf53c507