BACKUP YOUR DATABASE BEFORE UPDATING TO THIS REVISION!
This revision accounts for versioned sql updates to be stored in the {$schema}/upgrades directory.  Each file in the upgrades directory must be named using the numeric revision that it applies to with a ".sql" extension.  When the current database version is less that the one in version.php, all .sql files in the active schema's upgrades directory greater than the current revision will be executed.

In addition, any function named "upgrade_{revision number}" in the connection class for that schema will be executed if the revision number is greater than the current database revision.

The 1377.sql file in the schema/mysql/upgrades will convert existing databases to the utf-8 characters set and also convert text fields back into text fields upon completion (otherwise, they are turned into blob fields due to the binary conversion).

This should fix #201.  Principal code and sql upgrade comes from dbupgrade branch.  Thanks, gsnedders.

git-svn-id: http://svn.habariproject.org/habari/trunk@1530 653ae4dd-d31e-0410-96ef-6bf7bf53c507