Implementing a way to check for multiple PHP extensions easily.
Simply add an extension to the array $required_extensions and it will be checked. Use the key for the name, and the value for the URL to installation steps (php.net).

Also did some spelling corrections to "writeable". That part is not working yet, or has been removed, will investigate.

This fixes ticket #262, ticket #261 and ticket #259.
There is no way to backtrace an undefined variable to a specific module, all we can do is check if it's loaded before installation.

git-svn-id: http://svn.habariproject.org/habari/trunk@1564 653ae4dd-d31e-0410-96ef-6bf7bf53c507