This changeset fixes Issue #375:  "Dashboard says 1 Tag, but Tags Admin shows No Tags"
The problem was that tags could exist that were not related
to any post (for instance, the "habari" tag that the default
installation creates.  However, the queries used in both
retrieving all the tags in the tag set and in grabbing a single
tag for change or deletion used an INNER JOIN to the tag2post
table, resulting in the inability to remove or see tags not
related to a post but existing in the database.

M    admin/tag_collection.php
M    classes/tags.php
M    classes/adminhandler.php


git-svn-id: http://svn.habariproject.org/habari/trunk@1771 653ae4dd-d31e-0410-96ef-6bf7bf53c507