Adds the ability to add dashboard modules through the 'add item' module. There is a fair amount of work left to be done here, including creating a mechanism for plugins to add to the available modules list, making the add item module submit via AJAX so that you don't need to refresh the entire page, and updating the fetch_dash_module_* functions to be able to take an ID to deal with duplicate modules.
WARNING: This update will break any plugins that create dashboard widgets
because it changes the structure of the $modules array which is sent to the
dashboard template (Defensio is affected, probably others too). Disable any
such plugins if you want to be able to see the dashboard after doing 'svn up'.

Another note: there is a known bug in the code right now which causes added
modules to only appear after the dashboard has been reloaded once. This is
because of the order of events, where the add item form is processed only
after the module list is generated. This will be fixed when the 'add item'
form submits via AJAX.



git-svn-id: http://svn.habariproject.org/habari/trunk@1928 653ae4dd-d31e-0410-96ef-6bf7bf53c507