Upgrading to jQuery 1.3.2 and jQuery UI 1.7.1. The new jQuery UI slider unfortunately sets the slider position using percentages rather than pixels, and insists on having the max value be 100%. We *could* keep using it, but the code would be messy and indecipherable. So, I wrote a new slider.
This new slider mostly duplicates the old functionality. It does not yet
handle the case of a timeline which is larger than the view.  This feature
will be reimplemented soon.

Also, while this code works (at least on Firefox 3 and Safari 4), the slider
code really should be packaged as an instanceable object. I am committing this
working version before starting that.

Appologies in advance, because this commit also includes a coding standards
fix for parentheses after "if" statements.

Fixes #899.



git-svn-id: http://svn.habariproject.org/habari/trunk@3488 653ae4dd-d31e-0410-96ef-6bf7bf53c507