Set an actual value for the hidden control in the Habari silo's mkdir panel so the path to use during subdirectory creation is passed to where it's needed.
Related to this, also comment out the loading of the first subdirectory in media.js if there are no files in the root directory. This fixes ticket #971, but I'm not closing the ticket or removing these lines from the javascript yet because I have nothing to test with except the Habari silo. An empty panel doesn't look pretty, but broken functionality is even less pretty. If there are issues with the other silos, uncomment the lines and something else can be done.

git-svn-id: http://svn.habariproject.org/habari/trunk@3623 653ae4dd-d31e-0410-96ef-6bf7bf53c507